+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Release NEXT
 - Rewrite hook summary view using SummaryQuerySet.
 - Allow quotas to raise errors if their usage is over limit.
 - Fix monitoring_items serializer.
+- Fix filtering historic resources by customer.
 
 Release 0.100.0
 ---------------